+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.5.2]
+### Changed
+- Change `gen-utilities-data` to update target files without template.
+
 ## [1.5.1]
 ### Changed
-- Improved demo images on README.md
+- Improve demo images on README.md.
 
 ## [1.5.0]
 ### Changed